global: support disabling session activity tab
* Supports disabling session activity tab through configuration.
  (closes #201)
